feat : Factory 제거 및 확장 함수 추가
