+++ app/src/main/java/com/donglab/screennameviewer/SampleApplication.kt
@@ -2,15 +2,15 @@ package com.donglab.screennameviewer
 
 import android.app.Application
 import com.donglab.screennameviewer.config.ScreenNameOverlayConfig
+import com.donglab.screennameviewer.config.ScreenNameViewerConfiguration
 import com.donglab.screennameviewer.config.ScreenNameViewerSetting
-import com.donglab.screennameviewer.factory.ScreenNameViewerFactory
 
 class SampleApplication : Application() {
 
     override fun onCreate() {
         super.onCreate()
 
-        ScreenNameViewerFactory.initialize(
+        ScreenNameViewerConfiguration.initialize(
             application = this,
             settings = ScreenNameViewerSetting(
                 debugModeCondition = { true },
