+++ app/src/main/java/com/donglab/screennameviewer/compose/ComposeTestActivity.kt
@@ -2,6 +2,7 @@ package com.donglab.screennameviewer.compose
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
+import androidx.activity.compose.LocalActivity
 import androidx.activity.compose.setContent
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.getValue
@@ -14,9 +15,6 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
-import com.donglab.screennameviewer.config.ScreenNameOverlayConfig
-import com.donglab.screennameviewer.config.ScreenNameViewerSetting
-import com.donglab.screennameviewer.factory.ScreenNameViewerFactory
 import com.donglab.screennameviewer.compose.sample.screens.DetailScreen
 import com.donglab.screennameviewer.compose.sample.screens.HomeScreen
 import com.donglab.screennameviewer.compose.sample.screens.ProfileScreen
@@ -25,6 +23,7 @@ import com.donglab.screennameviewer.compose.sample.screens.dialogs.AboutDialog
 import com.donglab.screennameviewer.compose.sample.screens.dialogs.EditProfileDialog
 import com.donglab.screennameviewer.compose.sample.screens.nested.NestedScreen
 import com.donglab.screennameviewer.compose.sample.screens.tabs.TabScreen
+import com.donglab.screennameviewer.extensions.createComposeScreenNameViewer
 
 /**
  * Comprehensive test Activity for the enhanced Compose screen tracking system.
@@ -60,18 +59,14 @@ class ComposeTestActivity : ComponentActivity() {
 @androidx.compose.runtime.Composable
 private fun ComposeTestApp() {
     val navController = rememberNavController()
-    
-    val context = LocalContext.current
+    val activity = LocalActivity.current as? ComponentActivity
     
     // Navigation Screen Tracker 설정
     DisposableEffect(navController) {
-        val tracker = ScreenNameViewerFactory.createForCompose(
-            activity = context as ComponentActivity,
-            navController = navController
-        )
+        val screenNameViewer = activity?.createComposeScreenNameViewer(navController)
 
         onDispose {
-            tracker.cleanup()
+            screenNameViewer?.cleanup()
         }
     }
