+++ app/src/main/java/com/donglab/screennameviewer/compose/sample/fragments/PureComposeFragment.kt
@@ -6,6 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.activity.ComponentActivity
+import androidx.activity.compose.LocalActivity
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -17,14 +18,11 @@ import androidx.compose.runtime.DisposableEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.donglab.screennameviewer.config.ScreenNameOverlayConfig
-import com.donglab.screennameviewer.config.ScreenNameViewerSetting
-import com.donglab.screennameviewer.factory.ScreenNameViewerFactory
+import com.donglab.screennameviewer.extensions.createComposeScreenNameViewer
 
 /**
  * 순수 Compose로 구성된 Fragment (Navigation 포함)
@@ -49,17 +47,14 @@ class PureComposeFragment : Fragment() {
 @Composable
 private fun PureComposeContent() {
     val navController = rememberNavController()
-    val context = LocalContext.current
-    
-    // Navigation Screen Tracker 설정 (Fragment 내에서)
+    val activity = LocalActivity.current as? ComponentActivity
+
+    // Navigation Screen Tracker 설정
     DisposableEffect(navController) {
-        val tracker = ScreenNameViewerFactory.createForCompose(
-            activity = context as ComponentActivity,
-            navController = navController,
-        )
-        
+        val screenNameViewer = activity?.createComposeScreenNameViewer(navController)
+
         onDispose {
-            tracker.cleanup()
+            screenNameViewer?.cleanup()
         }
     }
